+++ apps/web-customer/src/core/Schema/OTPForm/validation.ts
@@ -1,13 +1,11 @@
 import * as yup from "yup";
-import { Regex } from "core-library/utils/regex";
-
-const { OTP_Regex } = Regex();
+import { OTP_REGEX } from "core-library/utils/regex";
 
 export const OTPSchema = yup.object().shape({
   otp: yup
     .string()
     .required("OTP is required")
-    .matches(OTP_Regex, "OTP must be exactly 6 digits"),
+    .matches(OTP_REGEX, "OTP must be exactly 6 digits"),
 });
 
 export type OTPType = yup.InferType<typeof OTPSchema>;
