+++ apps/web-customer/src/pages/account/change-password.tsx
@@ -1,0 +1,5 @@
+import { ChangePasswordBlock } from "../../components/blocks/ChangePasswordBlock/ChangePasswordBlock";
+
+export default function ChangePasswordPage() {
+  return <ChangePasswordBlock />;
+}
