+++ apps/web-customer/src/pages/account/verification/otp.tsx
@@ -1,0 +1,5 @@
+import OTPBlock from "../../../components/blocks/OTPBlock/OTPBlock";
+
+export default function OTPPage() {
+  return <OTPBlock />;
+}
