--- apps/web-customer/src/pages/changepassword.tsx
@@ -1,5 +1,0 @@
-import { ChangePasswordBlock } from "../components/blocks/ChangePasswordBlock/ChangePasswordBlock";
-
-export default function ChangePasswordPage() {
-  return <ChangePasswordBlock />;
-}
