--- apps/web-customer/src/pages/otp.tsx
@@ -1,5 +1,0 @@
-import OTPBlock from '../components/blocks/OTPBlock/OTPBlock';
-
-export default function OTPPage() {
-  return <OTPBlock />;
-}
