+++ packages/core-library/utils/Regex.ts
@@ -1,3 +1,3 @@
-export const Regex = () => ({
-  OTP_Regex: /^\d{6}$/,
-});
+export const OTP_REGEX = /^\d{6}$/;
+export const CONTAINS_NUMBER_REGEX = /(?=.*[0-9])/;
+export const CONTAINS_UPPERCASE_REGEX = /(?=.*[A-Z])/;
